+++ src/vs/workbench/contrib/notebook/browser/viewModel/baseCellViewModel.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Emitter, Event } from 'vs/base/common/event';
-import { Disposable, IDisposable, IReference } from 'vs/base/common/lifecycle';
+import { Disposable, dispose, IDisposable, IReference } from 'vs/base/common/lifecycle';
 import { Mimes } from 'vs/base/common/mime';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
@@ -23,7 +23,7 @@ import { CellViewModelStateChangeEvent } from 'vs/workbench/contrib/notebook/bro
 import { ViewContext } from 'vs/workbench/contrib/notebook/browser/viewModel/viewContext';
 import { NotebookCellTextModel } from 'vs/workbench/contrib/notebook/common/model/notebookCellTextModel';
 import { CellKind, INotebookCellStatusBarItem, INotebookSearchOptions } from 'vs/workbench/contrib/notebook/common/notebookCommon';
-import { NotebookOptionsChangeEvent } from 'vs/workbench/contrib/notebook/common/notebookOptions';
+import { getEditorTopPadding, NotebookOptionsChangeEvent } from 'vs/workbench/contrib/notebook/browser/notebookOptions';
 
 export abstract class BaseCellViewModel extends Disposable {
 
@@ -70,22 +70,6 @@ export abstract class BaseCellViewModel extends Disposable {
 
 	private _editState: CellEditState = CellEditState.Preview;
 
-	// get editState(): CellEditState {
-	// 	return this._editState;
-	// }
-
-	// set editState(newState: CellEditState) {
-	// 	if (newState === this._editState) {
-	// 		return;
-	// 	}
-
-	// 	this._editState = newState;
-	// 	this._onDidChangeState.fire({ editStateChanged: true });
-	// 	if (this._editState === CellEditState.Preview) {
-	// 		this.focusMode = CellFocusMode.Container;
-	// 	}
-	// }
-
 	private _lineNumbers: 'on' | 'off' | 'inherit' = 'inherit';
 	get lineNumbers(): 'on' | 'off' | 'inherit' {
 		return this._lineNumbers;
@@ -120,8 +104,8 @@ export abstract class BaseCellViewModel extends Disposable {
 	private _editorTransientState: IWordWrapTransientState | null = null;
 	private _resolvedCellDecorations = new Map<string, INotebookCellDecorationOptions>();
 
-	private readonly _cellDecorationsChanged = this._register(new Emitter<{ added: INotebookCellDecorationOptions[], removed: INotebookCellDecorationOptions[] }>());
-	onCellDecorationsChanged: Event<{ added: INotebookCellDecorationOptions[], removed: INotebookCellDecorationOptions[] }> = this._cellDecorationsChanged.event;
+	private readonly _cellDecorationsChanged = this._register(new Emitter<{ added: INotebookCellDecorationOptions[]; removed: INotebookCellDecorationOptions[] }>());
+	onCellDecorationsChanged: Event<{ added: INotebookCellDecorationOptions[]; removed: INotebookCellDecorationOptions[] }> = this._cellDecorationsChanged.event;
 
 	private _resolvedDecorations = new Map<string, {
 		id?: string;
@@ -149,6 +133,7 @@ export abstract class BaseCellViewModel extends Disposable {
 
 	set dragging(v: boolean) {
 		this._dragging = v;
+		this._onDidChangeState.fire({ dragStateChanged: true });
 	}
 
 	protected _textModelRef: IReference<IResolvedTextEditorModel> | undefined;
@@ -170,6 +155,8 @@ export abstract class BaseCellViewModel extends Disposable {
 		this._outputCollapsed = v;
 		this._onDidChangeState.fire({ outputCollapsedChanged: true });
 	}
+
+	private _isDisposed = false;
 
 	constructor(
 		readonly viewType: string,
@@ -213,7 +200,6 @@ export abstract class BaseCellViewModel extends Disposable {
 
 
 	abstract updateOptions(e: NotebookOptionsChangeEvent): void;
-	abstract hasDynamicHeight(): boolean;
 	abstract getHeight(lineHeight: number): number;
 	abstract onDeselect(): void;
 	abstract layoutChange(change: any): void;
@@ -232,7 +218,7 @@ export abstract class BaseCellViewModel extends Disposable {
 	// 	}
 	// }
 
-	attachTextEditor(editor: ICodeEditor) {
+	attachTextEditor(editor: ICodeEditor, estimatedHasHorizontalScrolling?: boolean) {
 		if (!editor.hasModel()) {
 			throw new Error('Invalid editor: model is missing');
 		}
@@ -250,22 +236,38 @@ export abstract class BaseCellViewModel extends Disposable {
 
 		if (this._editorViewStates) {
 			this._restoreViewState(this._editorViewStates);
+		} else {
+			// If no real editor view state was persisted, restore a default state.
+			// This forces the editor to measure its content width immediately.
+			if (estimatedHasHorizontalScrolling) {
+				this._restoreViewState({
+					contributionsState: {},
+					cursorState: [],
+					viewState: {
+						scrollLeft: 0,
+						firstPosition: { lineNumber: 1, column: 1 },
+						firstPositionDeltaTop: getEditorTopPadding()
+					}
+				});
+			}
 		}
 
 		if (this._editorTransientState) {
 			writeTransientState(editor.getModel(), this._editorTransientState, this._codeEditorService);
 		}
 
-		this._resolvedDecorations.forEach((value, key) => {
-			if (key.startsWith('_lazy_')) {
-				// lazy ones
-				const ret = this._textEditor!.deltaDecorations([], [value.options]);
-				this._resolvedDecorations.get(key)!.id = ret[0];
-			}
-			else {
-				const ret = this._textEditor!.deltaDecorations([], [value.options]);
-				this._resolvedDecorations.get(key)!.id = ret[0];
-			}
+		this._textEditor.changeDecorations((accessor) => {
+			this._resolvedDecorations.forEach((value, key) => {
+				if (key.startsWith('_lazy_')) {
+					// lazy ones
+					const ret = accessor.addDecoration(value.options.range, value.options.options);
+					this._resolvedDecorations.get(key)!.id = ret;
+				}
+				else {
+					const ret = accessor.addDecoration(value.options.range, value.options.options);
+					this._resolvedDecorations.get(key)!.id = ret;
+				}
+			});
 		});
 
 		this._editorListeners.push(this._textEditor.onDidChangeCursorSelection(() => { this._onDidChangeState.fire({ selectionChanged: true }); }));
@@ -278,16 +280,18 @@ export abstract class BaseCellViewModel extends Disposable {
 		this.saveViewState();
 		this.saveTransientState();
 		// decorations need to be cleared first as editors can be resued.
-		this._resolvedDecorations.forEach(value => {
-			const resolvedid = value.id;
+		this._textEditor?.changeDecorations((accessor) => {
+			this._resolvedDecorations.forEach(value => {
+				const resolvedid = value.id;
 
-			if (resolvedid) {
-				this._textEditor?.deltaDecorations([resolvedid], []);
-			}
+				if (resolvedid) {
+					accessor.removeDecoration(resolvedid);
+				}
+			});
 		});
 
 		this._textEditor = undefined;
-		this._editorListeners.forEach(e => e.dispose());
+		dispose(this._editorListeners);
 		this._editorListeners = [];
 		this._onDidChangeEditorAttachState.fire();
 
@@ -347,23 +351,28 @@ export abstract class BaseCellViewModel extends Disposable {
 			return decorationId;
 		}
 
-		const result = this._textEditor.deltaDecorations([], [decoration]);
-		this._resolvedDecorations.set(result[0], { id: result[0], options: decoration });
-		return result[0];
+		let id: string;
+		this._textEditor.changeDecorations((accessor) => {
+			id = accessor.addDecoration(decoration.range, decoration.options);
+			this._resolvedDecorations.set(id, { id, options: decoration });
+		});
+		return id!;
 	}
 
 	removeModelDecoration(decorationId: string) {
 		const realDecorationId = this._resolvedDecorations.get(decorationId);
 
 		if (this._textEditor && realDecorationId && realDecorationId.id !== undefined) {
-			this._textEditor.deltaDecorations([realDecorationId.id!], []);
+			this._textEditor.changeDecorations((accessor) => {
+				accessor.removeDecoration(realDecorationId.id!);
+			});
 		}
 
 		// lastly, remove all the cache
 		this._resolvedDecorations.delete(decorationId);
 	}
 
-	deltaModelDecorations(oldDecorations: string[], newDecorations: model.IModelDeltaDecoration[]): string[] {
+	deltaModelDecorations(oldDecorations: readonly string[], newDecorations: readonly model.IModelDeltaDecoration[]): string[] {
 		oldDecorations.forEach(id => {
 			this.removeModelDecoration(id);
 		});
@@ -417,7 +426,7 @@ export abstract class BaseCellViewModel extends Disposable {
 		return ret;
 	}
 
-	deltaCellStatusBarItems(oldItems: string[], newItems: INotebookCellStatusBarItem[]): string[] {
+	deltaCellStatusBarItems(oldItems: readonly string[], newItems: readonly INotebookCellStatusBarItem[]): string[] {
 		oldItems.forEach(id => {
 			const item = this._cellStatusBarItems.get(id);
 			if (item) {
@@ -556,6 +565,10 @@ export abstract class BaseCellViewModel extends Disposable {
 	async resolveTextModel(): Promise<model.ITextModel> {
 		if (!this._textModelRef || !this.textModel) {
 			this._textModelRef = await this._modelService.createModelReference(this.uri);
+			if (this._isDisposed) {
+				return this.textModel!;
+			}
+
 			if (!this._textModelRef) {
 				throw new Error(`Cannot resolve text model for ${this.uri}`);
 			}
@@ -596,14 +609,19 @@ export abstract class BaseCellViewModel extends Disposable {
 	}
 
 	override dispose() {
+		this._isDisposed = true;
 		super.dispose();
 
-		this._editorListeners.forEach(e => e.dispose());
-		this._undoRedoService.removeElements(this.uri);
+		dispose(this._editorListeners);
 
-		if (this._textModelRef) {
-			this._textModelRef.dispose();
+		// Only remove the undo redo stack if we map this cell uri to itself
+		// If we are not in perCell mode, it will map to the full NotebookDocument and
+		// we don't want to remove that entire document undo / redo stack when a cell is deleted
+		if (this._undoRedoService.getUriComparisonKey(this.uri) === this.uri.toString()) {
+			this._undoRedoService.removeElements(this.uri);
 		}
+
+		this._textModelRef?.dispose();
 	}
 
 	toJSON(): object {
